+++ src/interpreter/types/virtualMachine.ts
@@ -1,15 +1,28 @@
-import { type Node } from '../../ast/types';
-import { type SymbolTable } from './symbolTable';
 import { type Instr } from './instruction';
+import { type Memory } from '../../memory/memory';
+import { type Stack } from '../../utils/stack';
 
-export type CompilerMapping = {
-  [NodeType in Node['type']]: (
-    node: Extract<Node, { type: NodeType }>,
-    state: CompilerState
+export type VirtualMachineMapping = {
+  [InstrType in Instr['type']]: (
+    command: Extract<Instr, { type: InstrType }>,
+    state: VirtualMachineState
   ) => void;
 };
 
-export interface CompilerState {
-  instructions: Instr[];
-  symbolTable: SymbolTable;
+export interface VirtualMachineState {
+  memory: Memory;
+  stash: Stack<Value>;
 }
+
+export type Value = any;
+
+export interface Finished {
+  status: 'finished';
+  value: Value;
+}
+
+export interface Error {
+  status: 'error';
+}
+
+export type Result = Finished | Error;
