+++ src/interpreter/types/virtualMachineUtils.ts
