+++ src/interpreter/virtualMachine.ts
@@ -1,293 +1,133 @@
 import {
-  type CompilerMapping,
-  type CompilerState
+  type VirtualMachineMapping,
+  type VirtualMachineState,
+  type Result,
+  type Value
 } from './types/virtualMachine';
 import {
-  type ArrayAccessExpression,
-  type AssignmentExpression,
-  type BinaryExpression,
-  type BlockStatement,
-  type BreakStatement,
-  type CallExpression,
-  type ConditionalExpression,
-  type Constant,
-  type ContinueStatement,
-  type DefaultStatement,
-  type DoWhileStatement,
-  type EmptyStatement,
-  type ExpressionStatement,
-  type ForStatement,
-  type FunctionDeclaration,
-  type GotoStatement,
-  type Identifier,
-  type IdentifierStatement,
-  type IfStatement,
-  type LogicalExpression,
-  type MemberExpression,
-  type Node,
-  type Program,
-  type ReturnStatement,
-  type SequenceExpression,
-  type StringLiteral,
-  type SwitchStatement,
-  type UnaryExpression,
-  type UpdateExpression,
-  type VariableDeclaration,
-  type WhileStatement
-} from '../ast/types';
+  type AssignInstr,
+  type BinaryOperationInstr,
+  type CallInstr,
+  type DoneInstr,
+  type EnterProgramInstr,
+  type GotoInstr,
+  type Instr,
+  type JumpOnFalseInstr,
+  type LoadConstantInstr,
+  type LoadFunctionInstr,
+  type LoadSymbolInstr,
+  type TeardownInstr
+} from './types/instruction';
 import {
-  constructAssignInstr,
-  constructBinaryOperationInstr,
-  constructCallInstr,
-  constructDoneInstr,
-  constructEnterProgramInstr,
-  constructGotoInstr,
-  constructJumpOnFalseInstr,
-  constructLoadConstantInstr,
-  constructLoadFunctionInstr,
-  constructLoadSymbolInstr,
-  constructTeardownInstr,
-  PLACEHOLDER_ADDRESS
-} from './instruction';
-import { isEmptyStatement, isIdentifier } from '../ast/typeGuards';
-import { isNotUndefined } from '../utils/typeGuards';
-import {
-  InvalidCallError,
-  UnsupportedOperatorError,
-  UnsupportedOperatorErrorType
-} from './errors';
-import {
-  constructConditionalExpression,
-  constructFalseConstant,
-  constructMainCallExpression,
-  constructTrueConstant
-} from '../ast/constructors';
-import { type SymbolTable } from './types/symbolTable';
-import {
-  addBlockSymbolTableEntries,
-  addFunctionSymbolTableEntries,
-  addProgramSymbolTableEntries,
-  getFunctionSymbolTableEntry,
-  getNumOfEntriesInFrame,
-  getSymbolTableEntry,
-  getSymbolTableEntryInFrame
-} from './symbolTable';
-import { type Instr } from './types/instruction';
+  convertToAddress,
+  convertToPredicate,
+  evaluateBinaryExpression,
+  isTrue,
+  typeCheckBinaryOperation
+} from './virtualMachineUtils';
+import { Stack } from '../utils/stack';
+import { type Program } from '../ast/types';
+import { compileProgram } from './compiler';
+import { Memory } from '../memory/memory';
 
-export const compileProgram = (ast: Program): Instr[] => {
-  const symbolTable: SymbolTable = {
-    head: {},
-    tail: null,
-    parent: null
-  };
-  const state: CompilerState = {
-    symbolTable,
-    instructions: []
-  };
-  compile(ast, state);
-  const mainCallExpression = constructMainCallExpression();
-  compile(mainCallExpression, state);
-  const doneInstr = constructDoneInstr();
-  state.instructions.push(doneInstr);
-  return state.instructions;
+/**
+ * Evaluates the abstract syntax tree using a virtual machine evaluator &
+ * returns the result of evaluation asynchronously.
+ *
+ * @param ast The abstract syntax tree to evaluate.
+ */
+export const evaluate = async (ast: Program): Promise<Result> => {
+  return await new Promise(
+    (
+      resolve: (value: Result | PromiseLike<Result>) => void,
+      _reject: (reason?: any) => void
+    ) => {
+      try {
+        const instructions = compileProgram(ast);
+        const value = interpret(instructions);
+        resolve({ status: 'finished', value });
+      } catch (err) {
+        resolve({ status: 'error' });
+      }
+    }
+  );
 };
 
-const compile = (node: Node, state: CompilerState): void => {
-  // The typecast allows for mapping to a specific evaluator instr type from their union type.
-  // https://stackoverflow.com/questions/64527150/in-typescript-how-to-select-a-type-from-a-union-using-a-literal-type-property
-  compilers[node.type](node as any, state);
+export const interpret = (instructions: Instr[]): Value => {
+  const memory = new Memory(instructions, 1000, 1000, 1000);
+  const stash = new Stack<Value>();
+  const state: VirtualMachineState = {
+    memory,
+    stash
+  };
+  while (!state.memory.isAtDoneInstr()) {
+    const instr = state.memory.getCurrentInstr();
+    // The typecast allows for mapping to a specific evaluator instr type from their union type.
+    // https://stackoverflow.com/questions/64527150/in-typescript-how-to-select-a-type-from-a-union-using-a-literal-type-property
+    virtualMachineEvaluators[instr.type](instr as any, state);
+  }
+  return state.stash.size() === 0 ? undefined : state.stash.peek();
 };
 
-const compilers: CompilerMapping = {
-  ArrayAccessExpression: (
-    node: ArrayAccessExpression,
-    state: CompilerState
-  ) => {},
-  AssignmentExpression: (
-    node: AssignmentExpression,
-    state: CompilerState
-  ) => {},
-  BinaryExpression: (node: BinaryExpression, state: CompilerState) => {
-    compile(node.left, state);
-    compile(node.right, state);
-    const binaryOperationInstr = constructBinaryOperationInstr(node.operator);
-    state.instructions.push(binaryOperationInstr);
+const virtualMachineEvaluators: VirtualMachineMapping = {
+  Assign: (instr: AssignInstr, state: VirtualMachineState) => {
+    // TODO: Add conversion method to convert various stash values to their respective number.
+    // Do this when types are supported.
+    state.memory.setByOffset(instr.scope, instr.offset, state.stash.pop());
+    state.memory.moveToNextInstr();
   },
-  BlockStatement: (node: BlockStatement, state: CompilerState) => {
-    const blockSymbolTable = addBlockSymbolTableEntries(
-      node,
-      state.symbolTable
-    );
-    node.items.forEach((item) => {
-      state.symbolTable = blockSymbolTable;
-      compile(item, state);
-    });
-  },
-  BreakStatement: (node: BreakStatement, state: CompilerState) => {},
-  CallExpression: (node: CallExpression, state: CompilerState) => {
-    if (!isIdentifier(node.callee)) {
-      throw new InvalidCallError('Cannot call non-identifier.');
-    }
-    const functionEntry = getFunctionSymbolTableEntry(
-      node.callee.name,
-      state.symbolTable
-    );
-    if (functionEntry.numOfParams !== node.arguments.length) {
-      throw new InvalidCallError(
-        `Function takes in ${functionEntry.numOfParams} arguments but ${node.arguments.length} arguments were passed in.`
-      );
-    }
-
-    compile(node.callee, state);
-    // Compile in reverse order so that last argument is lower in the stash,
-    // and the first argument is higher.
-    for (let i = node.arguments.length - 1; i >= 0; i--) {
-      compile(node.arguments[i], state);
-    }
-    const callInstr = constructCallInstr(
-      node.arguments.length,
-      functionEntry.numOfVariables
-    );
-    state.instructions.push(callInstr);
-  },
-  ConditionalExpression: (
-    node: ConditionalExpression,
-    state: CompilerState
+  BinaryOperation: (
+    instr: BinaryOperationInstr,
+    state: VirtualMachineState
   ) => {
-    compile(node.predicate, state);
-    const jumpOnFalseInstr = constructJumpOnFalseInstr(PLACEHOLDER_ADDRESS);
-    state.instructions.push(jumpOnFalseInstr);
-    compile(node.consequent, state);
-    const gotoInstr = constructGotoInstr(PLACEHOLDER_ADDRESS);
-    state.instructions.push(gotoInstr);
-    jumpOnFalseInstr.instrAddress = state.instructions.length;
-    compile(node.alternate, state);
-    gotoInstr.instrAddress = state.instructions.length;
+    const right = state.stash.pop();
+    const left = state.stash.pop();
+    typeCheckBinaryOperation(instr.operator, left, right);
+    state.stash.push(evaluateBinaryExpression(instr.operator, left, right));
+    state.memory.moveToNextInstr();
   },
-  Constant: (node: Constant, state: CompilerState) => {
-    const loadConstantInstr = constructLoadConstantInstr(node.value);
-    state.instructions.push(loadConstantInstr);
-  },
-  ContinueStatement: (node: ContinueStatement, state: CompilerState) => {},
-  DefaultStatement: (node: DefaultStatement, state: CompilerState) => {},
-  DoWhileStatement: (node: DoWhileStatement, state: CompilerState) => {},
-  EmptyStatement: (node: EmptyStatement, state: CompilerState) => {},
-  ExpressionStatement: (node: ExpressionStatement, state: CompilerState) => {},
-  ForStatement: (node: ForStatement, state: CompilerState) => {},
-  FunctionDeclaration: (node: FunctionDeclaration, state: CompilerState) => {
-    const loadFunctionInstr = constructLoadFunctionInstr(PLACEHOLDER_ADDRESS);
-    state.instructions.push(loadFunctionInstr);
-    const gotoInstr = constructGotoInstr(PLACEHOLDER_ADDRESS);
-    state.instructions.push(gotoInstr);
-
-    loadFunctionInstr.functionInstrAddress = state.instructions.length;
-
-    // TODO: Replace param declarations when parameter list is supported
-    const functionSymbolTable = addFunctionSymbolTableEntries(
-      [],
-      node,
-      state.symbolTable
-    );
-    if (!isEmptyStatement(node.body)) {
-      node.body.items.forEach((item) => {
-        state.symbolTable = functionSymbolTable;
-        compile(item, state);
-      });
+  Call: (instr: CallInstr, state: VirtualMachineState) => {
+    // First item popped from the stash should be the arg for the first param and so on.
+    const args: Value[] = [];
+    for (let i = 0; i < instr.numOfArgs; i++) {
+      args.push(state.stash.pop());
     }
-    const teardownInstr = constructTeardownInstr();
-    state.instructions.push(teardownInstr);
-
-    gotoInstr.instrAddress = state.instructions.length;
-
-    const assignInstr = constructAssignInstr(
-      getSymbolTableEntry(node.id.name, state.symbolTable)
-    );
-    state.instructions.push(assignInstr);
+    const functionInstrAddress = convertToAddress(state.stash.pop());
+    state.memory.stackFunctionCallAllocate(args, instr.numOfVars);
+    state.memory.moveToInstr(functionInstrAddress);
   },
-  GotoStatement: (node: GotoStatement, state: CompilerState) => {},
-  Identifier: (node: Identifier, state: CompilerState) => {
-    const loadSymbolInstr = constructLoadSymbolInstr(
-      getSymbolTableEntry(node.name, state.symbolTable)
-    );
-    state.instructions.push(loadSymbolInstr);
+  Done: (instr: DoneInstr, state: VirtualMachineState) => {},
+  EnterProgram: (instr: EnterProgramInstr, state: VirtualMachineState) => {
+    state.memory.dataAllocate(instr.numOfDeclarations);
+    state.memory.moveToNextInstr();
   },
-  IdentifierStatement: (node: IdentifierStatement, state: CompilerState) => {},
-  IfStatement: (node: IfStatement, state: CompilerState) => {},
-  LogicalExpression: (node: LogicalExpression, state: CompilerState) => {
-    let conditionalExpression;
-    if (node.operator === '&&') {
-      conditionalExpression = constructConditionalExpression(
-        node.left,
-        node.right,
-        constructFalseConstant()
-      );
+  Goto: (instr: GotoInstr, state: VirtualMachineState) => {
+    state.memory.moveToInstr(instr.instrAddress);
+  },
+  JumpOnFalse: (instr: JumpOnFalseInstr, state: VirtualMachineState) => {
+    const predicate = convertToPredicate(state.stash.pop());
+    if (isTrue(predicate)) {
+      state.memory.moveToNextInstr();
+    } else {
+      state.memory.moveToInstr(instr.instrAddress);
     }
-    if (node.operator === '||') {
-      conditionalExpression = constructConditionalExpression(
-        node.left,
-        constructTrueConstant(),
-        node.right
-      );
-    }
-    if (conditionalExpression === undefined) {
-      throw new UnsupportedOperatorError(
-        node.operator,
-        UnsupportedOperatorErrorType.LOGICAL
-      );
-    }
-    compile(conditionalExpression, state);
   },
-  MemberExpression: (node: MemberExpression, state: CompilerState) => {},
-  Program: (node: Program, state: CompilerState) => {
-    const programSymbolTable = addProgramSymbolTableEntries(
-      node,
-      state.symbolTable
-    );
-    const enterProgramInstr = constructEnterProgramInstr(
-      getNumOfEntriesInFrame(programSymbolTable.head)
-    );
-    state.instructions.push(enterProgramInstr);
-    node.body.forEach((item) => {
-      state.symbolTable = programSymbolTable;
-      compile(item, state);
-    });
+  LoadConstant: (instr: LoadConstantInstr, state: VirtualMachineState) => {
+    state.stash.push(instr.value);
+    state.memory.moveToNextInstr();
   },
-  ReturnStatement: (node: ReturnStatement, state: CompilerState) => {
-    // TODO: Check if return with no argument works correctly.
-    if (isNotUndefined(node.argument)) {
-      if (node.argument.expressions.length > 1) {
-        throw new InvalidCallError('Encountered more than 1 return value');
-      }
-      compile(node.argument.expressions[0], state);
-    }
-    // TODO: Handle tail call.
-    const teardownInstr = constructTeardownInstr();
-    state.instructions.push(teardownInstr);
+  LoadFunction: (instr: LoadFunctionInstr, state: VirtualMachineState) => {
+    state.stash.push(instr.functionInstrAddress);
+    state.memory.moveToNextInstr();
   },
-  SequenceExpression: (node: SequenceExpression, state: CompilerState) => {
-    node.expressions.forEach((expression) => {
-      compile(expression, state);
-    });
+  LoadSymbol: (instr: LoadSymbolInstr, state: VirtualMachineState) => {
+    const value = state.memory.getByOffset(instr.scope, instr.offset);
+    state.stash.push(value);
+    state.memory.moveToNextInstr();
   },
-  StringLiteral: (node: StringLiteral, state: CompilerState) => {},
-  SwitchStatement: (node: SwitchStatement, state: CompilerState) => {},
-  UnaryExpression: (node: UnaryExpression, state: CompilerState) => {},
-  UpdateExpression: (node: UpdateExpression, state: CompilerState) => {},
-  VariableDeclaration: (node: VariableDeclaration, state: CompilerState) => {
-    node.declarations.forEach((declarator) => {
-      const initialValue = declarator.initialValue;
-      if (isNotUndefined(initialValue)) {
-        compile(initialValue, state);
-        // Declaration names should have been added to the symbol table by the parent scope.
-        // Should only need to assign for declaration in last frame.
-        const entry = getSymbolTableEntryInFrame(
-          declarator.id.name,
-          state.symbolTable.head
-        );
-        const assignInstr = constructAssignInstr(entry);
-        state.instructions.push(assignInstr);
-      }
-    });
-  },
-  WhileStatement: (node: WhileStatement, state: CompilerState) => {}
+  Teardown: (instr: TeardownInstr, state: VirtualMachineState) => {
+    const returnAddress = state.memory.getReturnAddress();
+    state.memory.stackFunctionCallTeardown();
+    state.memory.moveToInstr(returnAddress);
+  }
 };
